requirements.txt
flask==2.2.3
PyYAML==6.0
pyfiglet==0.8.post1
termcolor==2.2.0
click==8.1.3
click_log==0.4.0
flask-restx==0.2.0
requests==2.24.0
uvicorn==0.20.0
numpy==1.24.2
pandas==1.5.3
tqdm==4.65.0